requirements.txt
streamlit
pycryptodome
numpy
pandas
tensorflow-intel
cryptography
